git-tfs-id: [http://devvm01tfs03.telamericamedia.com:8080/tfs/CMWPA]$/Telamerica/Broadcast/OnDeck;C20722

+++ Source/Tam/Maestro/Services/Broadcast/Services.Broadcast/Entities/StationProgramAudienceRateDto.cs
@@ -13,8 +13,8 @@ namespace Services.Broadcast.Entities
         public List<LookupDto> Genres { get; set; }
         public decimal? Rate15 { get; set; }
         public decimal? Rate30 { get; set; }
-        public float? Impressions { get; set; }
-        public float? Rating { get; set; }
+        public double? Impressions { get; set; }
+        public double? Rating { get; set; }
         public List<FlightWeekDto> Flights { get; set; }
         public string Flight { get; set; }
         public DateTime FlightStartDate { get; set; }
